+++ aws/container-linux/kubernetes/variables.tf
@@ -44,7 +44,7 @@ variable "worker_type" {
 variable "os_image" {
   type        = string
   description = "AMI channel for a Container Linux derivative (coreos-stable, coreos-beta, coreos-alpha, flatcar-stable, flatcar-beta, flatcar-alpha, flatcar-edge)"
-  default     = "coreos-stable"
+  default     = "flatcar-stable"
 }
 
 variable "disk_size" {
@@ -94,12 +94,6 @@ variable "worker_snippets" {
 variable "ssh_authorized_key" {
   type        = string
   description = "SSH public key for user 'core'"
-}
-
-variable "asset_dir" {
-  type        = string
-  description = "Absolute path to a directory where generated assets should be placed (contains secrets)"
-  default     = ""
 }
 
 variable "networking" {
@@ -154,6 +148,12 @@ variable "worker_node_labels" {
 }
 
 # unofficial, undocumented, unsupported
+
+variable "asset_dir" {
+  type        = string
+  description = "Absolute path to a directory where generated assets should be placed (contains secrets)"
+  default     = ""
+}
 
 variable "cluster_domain_suffix" {
   type        = string