+++ aws/fedora-coreos/kubernetes/variables.tf
@@ -44,7 +44,7 @@ variable "worker_type" {
 variable "os_image" {
   type        = string
   description = "AMI channel for Fedora CoreOS (not yet used)"
-  default     = "coreos-stable"
+  default     = "stable"
 }
 
 variable "disk_size" {