+++ google-cloud/container-linux/kubernetes/variables.tf
@@ -48,8 +48,7 @@ variable "worker_type" {
 
 variable "os_image" {
   type        = string
-  description = "Container Linux image for compute instances (e.g. coreos-stable)"
-  default     = "coreos-stable"
+  description = "Container Linux image for compute instances (e.g. coreos-stable, custom-image)"
 }
 
 variable "disk_size" {
@@ -81,12 +80,6 @@ variable "worker_snippets" {
 variable "ssh_authorized_key" {
   type        = string
   description = "SSH public key for user 'core'"
-}
-
-variable "asset_dir" {
-  type        = string
-  description = "Absolute path to a directory where generated assets should be placed (contains secrets)"
-  default     = ""
 }
 
 variable "networking" {
@@ -130,6 +123,12 @@ variable "worker_node_labels" {
 }
 
 # unofficial, undocumented, unsupported
+
+variable "asset_dir" {
+  type        = string
+  description = "Absolute path to a directory where generated assets should be placed (contains secrets)"
+  default     = ""
+}
 
 variable "cluster_domain_suffix" {
   type        = string